require post parameters for TeamPerformTestView
This sentry error:
https://sentry.ccnmtl.columbia.edu/sentry-internal/brownfield_django/group/993/

It appears that the flash component is sometimes sending back invalid
requests, with 'undefined' instead of an integer value for some
parameters.

Ultimately, that's a bug in the flash part and ought to be fixed
there. But that's not going to happen any time soon. In the meantime,
the correct behavior for the backend is to respond with a 400 bad
request rather than throwing up a 500.

So I pulled out some basic validation. It's checking the parameters that
it's obviously expecting. I left a few off like 'description' and
'testDetails' because it's not clear to me whether those are actually
required or if it's OK for them to be None...