Resolve 500 error
AnonymousUser hitting the /ccnmtl/home/ url (CCNMTLHomeView) should be redirected to /login/, instead a 500 error results. The LoggedIn mixins are not reached as the view overrides dispatch and attempts to access attributes on request.user that do not exist on AnonymousUser. This PR creates a new mixin to verify a logged in user's profile id matches the passed parameter. The mixin is then called after the two LoggedInMixins.