add missing else
it was just bugging me to see an if/elif/elif with no else case in
there. it leaves this implicit case hanging there with no indication
whether that's a valid case or not (it's not).

I reversed the order so 'student' is the default fallback. If `role()`
is used for any sort of auth, that ensures that a bug in there would
most likely err on the side of the user getting a lower role than they
should rather than accidental privelege escalation.