extract method admin_or_team_url()
the logic here is a bit twisted. If the user is in the admin group, a
profile is created and they get redirected to '/' otherwise it looks for
a team for the user and redirects them to that page. I've maintained
that behavior, though it feels overly complicated to me (and is wrapped
in a bare `except` and has no tests...)