Add document download links again - PMT #109570
I had previously added these links, and it caused the error noted at
PMT #109671. I've tested this new change on staging, but we will want to
test on production as well. We can do this by logging in as ccnmtl and
clicking the "Demo" tab, to make sure everything appears.

The static link needs to not include the initial slash, like:
`{% static 'flash/documents/slCatalog.pdf' %}`

And not:
`{% static '/flash/documents/slCatalog.pdf' %}`